+++ cli/ferrule_cli/__init__.py
@@ -1,0 +1,1 @@
+"""Ferrule command-line interface."""